added TransitionTimer,renamed activity, added pagage Stats
