--- app/src/main/java/de/hsaugsburg/teamulster/sohappy/stateMachine/State.kt
@@ -1,6 +1,0 @@
-package de.hsaugsburg.teamulster.sohappy.stateMachine
-
-interface State {
-    fun consumeAction(action: Action): State
-    fun activity()
-}
