+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/stateMachine/states/NoSmile.kt
@@ -1,4 +1,6 @@
-package de.hsaugsburg.teamulster.sohappy.stateMachine
+package de.hsaugsburg.teamulster.sohappy.stateMachine.states
+
+import de.hsaugsburg.teamulster.sohappy.stateMachine.Action
 
 class NoSmile : State {
     override fun consumeAction(action: Action): State {
@@ -9,6 +11,6 @@ class NoSmile : State {
         }
     }
 
-    override fun activity() {
+    override fun executeCoreFunctionality() {
     }
 }
