+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/stateMachine/states/Questions.kt
@@ -1,4 +1,6 @@
-package de.hsaugsburg.teamulster.sohappy.stateMachine
+package de.hsaugsburg.teamulster.sohappy.stateMachine.states
+
+import de.hsaugsburg.teamulster.sohappy.stateMachine.Action
 
 class Questions : State {
     override fun consumeAction(action: Action): State {
@@ -8,6 +10,6 @@ class Questions : State {
         }
     }
 
-    override fun activity() {
+    override fun executeCoreFunctionality() {
     }
 }
