+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/stateMachine/states/State.kt
@@ -1,0 +1,8 @@
+package de.hsaugsburg.teamulster.sohappy.stateMachine.states
+
+import de.hsaugsburg.teamulster.sohappy.stateMachine.Action
+
+interface State {
+    fun consumeAction(action: Action): State
+    fun executeCoreFunctionality()
+}
