+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/stateMachine/states/WaitingForFace.kt
@@ -1,4 +1,6 @@
-package de.hsaugsburg.teamulster.sohappy.stateMachine
+package de.hsaugsburg.teamulster.sohappy.stateMachine.states
+
+import de.hsaugsburg.teamulster.sohappy.stateMachine.Action
 
 class WaitingForFace : State {
 
@@ -10,7 +12,7 @@ class WaitingForFace : State {
         }
     }
 
-    override fun activity() {
+    override fun executeCoreFunctionality() {
         startFaceDetection()
     }
